+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/ContainerParameters.java
@@ -18,6 +18,8 @@ package com.google.cloud.tools.jib.gradle;
 
 import com.google.cloud.tools.jib.image.ImageFormat;
 import com.google.cloud.tools.jib.image.json.BuildableManifestTemplate;
+import com.google.cloud.tools.jib.plugins.common.ConfigurationPropertyValidator;
+import com.google.cloud.tools.jib.plugins.common.PropertyNames;
 import com.google.common.base.Preconditions;
 import java.util.Collections;
 import java.util.List;
@@ -46,6 +48,9 @@ public class ContainerParameters {
   @Input
   @Optional
   public boolean getUseCurrentTimestamp() {
+    if (System.getProperty(PropertyNames.CONTAINER_USE_CURRENT_TIMESTAMP) != null) {
+      return Boolean.getBoolean(PropertyNames.CONTAINER_USE_CURRENT_TIMESTAMP);
+    }
     return useCurrentTimestamp;
   }
 
@@ -56,6 +61,10 @@ public class ContainerParameters {
   @Input
   @Optional
   public List<String> getEntrypoint() {
+    if (System.getProperty(PropertyNames.CONTAINER_ENTRYPOINT) != null) {
+      return ConfigurationPropertyValidator.parseListProperty(
+          System.getProperty(PropertyNames.CONTAINER_ENTRYPOINT));
+    }
     return entrypoint;
   }
 
@@ -66,6 +75,10 @@ public class ContainerParameters {
   @Input
   @Optional
   public List<String> getJvmFlags() {
+    if (System.getProperty(PropertyNames.CONTAINER_JVM_FLAGS) != null) {
+      return ConfigurationPropertyValidator.parseListProperty(
+          System.getProperty(PropertyNames.CONTAINER_JVM_FLAGS));
+    }
     return jvmFlags;
   }
 
@@ -76,6 +89,10 @@ public class ContainerParameters {
   @Input
   @Optional
   public Map<String, String> getEnvironment() {
+    if (System.getProperty(PropertyNames.CONTAINER_ENVIRONMENT) != null) {
+      return ConfigurationPropertyValidator.parseMapProperty(
+          System.getProperty(PropertyNames.CONTAINER_ENVIRONMENT));
+    }
     return environment;
   }
 
@@ -87,6 +104,9 @@ public class ContainerParameters {
   @Nullable
   @Optional
   public String getMainClass() {
+    if (System.getProperty(PropertyNames.CONTAINER_MAIN_CLASS) != null) {
+      return System.getProperty(PropertyNames.CONTAINER_MAIN_CLASS);
+    }
     return mainClass;
   }
 
@@ -97,6 +117,10 @@ public class ContainerParameters {
   @Input
   @Optional
   public List<String> getArgs() {
+    if (System.getProperty(PropertyNames.CONTAINER_ARGS) != null) {
+      return ConfigurationPropertyValidator.parseListProperty(
+          System.getProperty(PropertyNames.CONTAINER_ARGS));
+    }
     return args;
   }
 
@@ -107,6 +131,10 @@ public class ContainerParameters {
   @Input
   @Optional
   public Class<? extends BuildableManifestTemplate> getFormat() {
+    if (System.getProperty(PropertyNames.CONTAINER_FORMAT) != null) {
+      return ImageFormat.valueOf(System.getProperty(PropertyNames.CONTAINER_FORMAT))
+          .getManifestTemplateClass();
+    }
     return Preconditions.checkNotNull(format).getManifestTemplateClass();
   }
 
@@ -117,6 +145,10 @@ public class ContainerParameters {
   @Input
   @Optional
   public List<String> getPorts() {
+    if (System.getProperty(PropertyNames.CONTAINER_PORTS) != null) {
+      return ConfigurationPropertyValidator.parseListProperty(
+          System.getProperty(PropertyNames.CONTAINER_PORTS));
+    }
     return ports;
   }
 
@@ -127,6 +159,10 @@ public class ContainerParameters {
   @Input
   @Optional
   public Map<String, String> getLabels() {
+    if (System.getProperty(PropertyNames.CONTAINER_LABELS) != null) {
+      return ConfigurationPropertyValidator.parseMapProperty(
+          System.getProperty(PropertyNames.CONTAINER_LABELS));
+    }
     return labels;
   }
 
@@ -137,6 +173,9 @@ public class ContainerParameters {
   @Input
   @Optional
   public String getAppRoot() {
+    if (System.getProperty(PropertyNames.CONTAINER_APP_ROOT) != null) {
+      return System.getProperty(PropertyNames.CONTAINER_APP_ROOT);
+    }
     return appRoot;
   }